requirements.txt
pandas
fasttext
lxml
Flask==0.12
requests==2.9.1 
markovify==0.8.3
nltk==3.4.5
# torch==1.6.0
# torchvision==0.6.0
beautifulsoup4==4.6.0
pillow==8.2.0